Refactor scroll and effects components to dedicated directories
- Created scroll directory with subcomponents:
  - scroll-progress component
  - scroll-to-top component
  - section-anchor component
- Created effects directory with subcomponents:
  - parallax-section component
  - magnetic component
- Added comprehensive documentation with README files and usage examples
- Created consistent directory structure with index.ts exports
- Added backward compatibility re-exports with @deprecated annotations
- Updated COMPONENT-MOVEMENT-PLAN.md to track changes
