+++ src/components/ui/effects/index.ts
@@ -1,0 +1,2 @@
+export * from "./parallax";
+export * from "./magnetic";
