+++ src/components/ui/effects/magnetic/index.ts
@@ -1,0 +1,1 @@
+export * from "./magnetic";
