+++ src/components/ui/effects/parallax/index.ts
@@ -1,0 +1,1 @@
+export * from "./parallax-section";
