+++ src/components/ui/scroll/anchor/index.ts
@@ -1,0 +1,1 @@
+export * from "./section-anchor";
