+++ src/components/ui/scroll/index.ts
@@ -1,0 +1,3 @@
+export * from "./anchor";
+export * from "./progress";
+export * from "./to-top";
