+++ src/components/ui/scroll/progress/index.ts
@@ -1,0 +1,1 @@
+export * from "./scroll-progress";
