+++ src/components/ui/scroll/to-top/index.ts
@@ -1,0 +1,1 @@
+export * from "./scroll-to-top";
